### 2019-02-05 #1235 @whedon
[PRE REVIEW]: AtlasReader: A Python package to generate coordinate tables, region labels, and informative figures from statistical MRI images
**Submitting author:** @miykael (<a href="http://orcid.org/0000-0002-5866-047X">Michael Philipp Notter</a>)
**Repository:** <a href="https://github.com/miykael/atlasreader" target ="_blank">https://github.com/miykael/atlasreader</a>
**Version:** v0.0.3
