### 2019-02-05 9e1b887
FIX: update figure size and caption for publication
