### 2019-02-05 0832f8a
FIX: update figure size and caption
