### 2019-02-15 #1235 @cMadan
@andrewheusser, sooner is good, but it's also flexible. In all likelihood, you'll have some comments that will take the authors time to work on for a week or two, and then will need to look again and see if sufficiently revised. That back-and-forth can go on for just a week or two, or sometimes is more intermittent over a few months, it all depends how responsive both the authors and reviewer are.
