### 2019-02-15 #1235 @miykael
Hi @andrewheusser - thank you for reviewing this. And thank you for starting it so quickly. I will respond to your review and comments as quickly as possible.

@cMadan might know of some additional guide/checklist, but I think the main one from JOSS is here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
