### 2019-02-19 bc94b31
WIP: update travis matrix version definition
