### 2019-02-19 99fa3c4
WIP: correct travis matrix definition
