### 2019-02-19 460ed45
FIX: correct python version definition
