+++ .travis.yml
@@ -1,10 +1,6 @@
 language: python
 sudo: false
 
-python:
-    - 3.6
-    - 3.7
-
 matrix:
   include:
     - python: 3.6
