### 2019-02-19 d261701
WIP: travis simplication plus adding dist xenial
