+++ .travis.yml
@@ -1,26 +1,15 @@
 language: python
 sudo: false
+dist: xenial
 
-matrix:
-  include:
-    - python: 3.6
-      env:
-        - STYLE=1
-    - python: 3.6
-      env:
-        - COVERAGE=1
-    - python: 3.6
-      env:
-        - NOTEBOOKS=1
-    - python: 3.7
-      env:
-        - STYLE=1
-    - python: 3.7
-      env:
-        - COVERAGE=1
-    - python: 3.7
-      env:
-        - NOTEBOOKS=1
+python:
+    - 3.6
+    - 3.7
+
+env:
+    - STYLE=1
+    - COVERAGE=1
+    - NOTEBOOKS=1
 
 before_install:
     - python -m pip install --upgrade pip
