### 2019-02-19 ec11fa3
FIX: correct peak table checksum
