### 2019-02-19 e012db4
FIX: correct cluster table checksum
