### 2019-02-19 8705774
ENH: add mybinder and nbviewer links to readme
