### 2019-02-20 0b99f4f
ENH: debug code and improve documentation of voxel_thresh
