### 2019-02-20 miykael/atlasreader#105 @andrewheusser
JOSS REVIEW: scipy and sklearn missing DOIs
If they are available, please add DOIs to the scipy and sklearn references
