### 2019-02-20 23cae1f
FIX: add default value to plot_kws description
