### 2019-02-20 8b42ce1
FIX: clearly separate atlasreader and create_output
