### 2019-02-20 3134231
FIX: integrate JOSS reviewer suggestion
