### 2019-02-20 b2628f4
FIX: add url to sklearn reference
