+++ paper/paper.bib
@@ -105,7 +105,8 @@
   volume={12},
   number={Oct},
   pages={2825--2830},
-  year={2011}
+  year={2011},
+  url="https://scikit-learn.org/"
 }
 @article{scikitimage,
   title={scikit-image: image processing in Python},
