### 2019-02-20 miykael/atlasreader#105 @miykael
Surprisingly, both of them don't have a DOI. I've looked again and wasn't able to find one.
But I've added the sklearn URL to the reference (scipy had already one), with this commit https://github.com/miykael/atlasreader/pull/107/commits/b2628f4d4182343641b857a824afd4d0c870d22f.
