### 2019-02-20 4e277c2
ENH: add thresholding 'direction' parameter
