### 2019-02-21 f59357e
ENH: check output table against expected numpy array
