### 2019-02-21 31d2406
FIX: reduce test time with less cluster figures
