### 2019-02-21 b2fdb8d
FIX: correct PEP8 style issues
