+++ atlasreader/tests/test_atlasreader.py
@@ -42,19 +42,20 @@ EXAMPLE_COORDS = dict(
     bounding_shape=np.array([90, 90, 90])
 )
 EXPECTED_TABLES = dict(
-    cluster=np.array([[ 42, -25,  58,  6.66003, 36936],
+    cluster=np.array([[42,  -25,  58,  6.66003, 36936],
                       [-36, -25,  55, -6.63604, 15012],
-                      [ 45, -19,  16,  5.76538,  7722],
+                      [45,  -19,  16,  5.76538,  7722],
                       [-15, -52, -26,  6.11673,  7101],
-                      [ 18, -55, -23, -5.90086,  5184],
+                      [18,  -55, -23, -5.90086,  5184],
                       [-36, -19,  19, -5.00687,   648]]),
-    peak=np.array([[ 42, -25,  58,  7.94135, 36936],
+    peak=np.array([[42,  -25,  58,  7.94135, 36936],
                    [-36, -25,  55, -7.94144, 15012],
-                   [ 45, -19,  16,  7.94135,  7722],
+                   [45,  -19,  16,  7.94135,  7722],
                    [-15, -52, -26,  7.94135,  7101],
-                   [ 18, -55, -23, -7.94144,  5184],
+                   [18,  -55, -23, -7.94144,  5184],
                    [-36, -19,  19, -6.21808,   648]])
 )
+
 
 def test_get_atlases():
     for atlas in atlasreader._ATLASES:
