### 2019-02-22 #1257 @andrewheusser
hey @cMadan - I finished the review and the author sufficiently addressed all of my concerns (see [here](https://github.com/miykael/atlasreader/issues?q=is%3Aissue+author%3Aandrewheusser+is%3Aclosed).  The author implemented the changes on a branch, but they are not yet merged.  What are the next steps?
