### 2019-02-22 #1257 @cMadan
@andrewheusser, thank you for the prompt and thorough review! It looks like you're all done.

For @miykael, the next steps are to merge these changes and confirm that you think it's ready to be accepted. Then I'll look everything over and check that I agree. If that's the case, I'll then ask for the DOI of an archived copy of the code and what version of the software corresponds to the code after the revisions related to the review--and then we're all set!
