+++ atlasreader/info.py
@@ -1,4 +1,4 @@
-__version__ = '0.0.4'
+__version__ = '0.1.0'
 
 NAME = 'atlasreader'
 MAINTAINER = 'Michael Notter'
