### 2019-02-22 #1257 @miykael
Thank you @andrewheusser for the very nice and helpful feedback! Your comments were much appreciated and helped to improve the toolbox.

@cMadan - I've merged the changes created a new GitHub release and a new PyPi release. So, the current version number is now v0.1.0. In my opinion (and this of the co-authors), the toolbox is ready to be accepted. Can I update the version number in the JOSS publication or is that up to you? Is there anything else that I can help with?
