### 2019-02-22 aba2e86
FIX: correct new line after figure/table description
