### 2019-02-22 2a1b4cc
ENH: update figure text and reformat their layout
