### 2019-02-22 df8621d
FIX: small changes to effect layout of final paper
