### 2019-02-22 #1257 @miykael
I apologize for the repeated `@whedon` commands, but I try to get the paper output to the right format. But the next "PDF generation" should be the last.
