### 2019-02-22 e95aa6b
FIX: correct small bug in references
