+++ paper/paper.bib
@@ -94,7 +94,7 @@
 @article{scipy,
   author     = {Eric Jones and Travis Oliphant and Pearu Peterson and others},
   title      = {{SciPy}: Open source scientific tools for {Python}},
-  year       = {2001--},
+  year       = {2001},
   url        = "http://www.scipy.org/",
   note       = {[Online; accessed 24. Sept. 2018]}
 }
