### 2019-02-22 #1257 @miykael
Sorry, found a small typo in the references. But otherwise, everything is good to go.
