### 2019-02-24 #1257 @cMadan
I think those missing DOIs aren't correct here, they are related but more tutorial to the packages than for the packages per se. As discussed in the review (https://github.com/miykael/atlasreader/issues/105), I think there aren't DOIs for those references.

@miykael, you're all set congratulations! @andrewheusser, thank you again for the thorough review.

@arfon, can you do the honors?
