### 2019-02-24 #1257 @arfon
> @arfon, can you do the honors?

Done. The proof I generated locally ended up doing a slightly better job with figure layout so I processed the paper manually.
